==> 3_match_data.ipynb <==
### 1. Import pandas for merging pricing file and review file

import pandas as pd

### 2. Read CSV files and merge it

pricing_df = pd.read_csv('data_pricing_apps.csv', sep=';').reset_index()
reviews_df = pd.read_csv('small_data_clean.csv', sep=',').reset_index()
df = pd.merge(pricing_df, reviews_df, how='inner', on='app_name')

df['monthly_price'] = df['monthly_price'].div(100)      #divide prices by 100 to get usual prices
df['yearly_price'] = df['yearly_price'].div(100)

### 3. Create bins for monthly prices and score

price_labels = ['1', '2', '3', '4']
price_bins = [0, 7.5, 15, 70, 300]
df['price_category'] = pd.cut(df['monthly_price'], bins=price_bins, labels=price_labels)

sentiment_labels = ['negative', 'neutral', 'positive']
sentiment_bins = [0, 2, 3, 5]
df['sentiment'] = pd.cut(df['score'], bins=sentiment_bins, labels=sentiment_labels)

### 4. Export merged file as CSV

df.to_csv('clean_data_+_data_pricing.csv')
